refactor: UserController 및 UserService 클래스에서 로그인 관련 기능 분리
