+++ src/test/java/techcourse/myblog/web/UserControllerTest.java
@@ -81,39 +81,4 @@ class UserControllerTest {
                             );
                 });
     }
-
-    @Test
-    public void 로그인_테스트() {
-        /** 회원 가입 **/
-        webTestClient.post()
-                .uri("/users")
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-                .body(BodyInserters
-                        .fromFormData("name", "코니코니")
-                        .with("email", "이메일")
-                        .with("password", "비번비번"))
-                .exchange()
-                .expectStatus().isFound()
-                .expectBody()
-                .consumeWith(response -> {
-                    /** 로그인 **/
-                    webTestClient.post()
-                            .uri("/login")
-                            .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-                            .body(BodyInserters
-                                    .fromFormData("email", "이메일이메일")
-                                    .with("password", "비번비번"))
-                            .exchange()
-                            .expectStatus().isFound()
-                            .expectBody()
-                            .consumeWith(r -> {
-                                /** 인덱스 페이지 이동 **/
-                                webTestClient.get()
-                                        .uri(r.getRequestHeaders().getLocation())
-                                        .exchange()
-                                        .expectStatus()
-                                        .isOk();
-                            });
-                });
-    }
 }
